Change index queue implementation to bulk from single.
Indexing queue now fetches a list of oids to index instead of a single oid to avoid race conditions
when adding new oids.

Deleting a bulk of oids required overwriting some elastisch-libarary code since the delete-by-query api calls the
wrong url and does nothing. At the time of this commit an issue was posted to https://github.com/clojurewerkz/elastisch/issues
and the overwritten methods should be fixed to use elastisch code once (if ever) they are fixed.